restructured extraction log according to Mapping props
