+++ packages/itwin/grouping-mapping-widget/src/widget/components/MappingImportWizardModal.tsx
@@ -9,7 +9,7 @@ import { Stepper } from "@itwin/itwinui-react";
 import { Modal } from "@itwin/itwinui-react";
 import React, { useEffect, useRef, useState } from "react";
 import ConfirmMappingImport from "./ConfirmMappingsImport";
-import type { IMappingTyped } from "./Mapping";
+import type { IMappingTyped } from "./Mappings";
 import "./MappingImportWizardModal.scss";
 import SelectIModel from "./SelectIModel";
 import SelectITwin from "./SelectITwin";
@@ -30,7 +30,7 @@ const defaultDisplayStrings = {
 };
 interface MappingImportWizardModalProps {
   show: boolean;
-  setShow: React.Dispatch<React.SetStateAction<boolean>>;
+  setShow: (show: boolean) => void;
   onFinish: () => Promise<void>;
   displayStrings?: Partial<typeof defaultDisplayStrings>;
 }
