+++ packages/itwin/grouping-mapping-widget/src/widget/components/MappingsView.scss
@@ -1,0 +1,37 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
+
+.gmw-mappings-view-container {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  min-height: 0;
+
+  .gmw-table-toolbar {
+    display: flex;
+    justify-content: space-between;
+    gap: var(--iui-size-xs);
+    padding-bottom: var(--iui-size-s);
+    flex-wrap: wrap;
+
+    .gmw-button-spacing {
+      display: flex;
+      gap: var(--iui-size-xs);
+    }
+  }
+
+  .gmw-mappings-border {
+    padding-top: var(--iui-size-s);
+    border-bottom: 1px solid var(--iui-color-border-subtle);
+  }
+
+  .gmw-mappings-list {
+    display: flex;
+    flex-direction: column;
+    overflow-y: overlay;
+    gap: var(--iui-size-xs);
+    margin-top: var(--iui-size-xs);
+  }
+}
